pyproject.toml
[project]
name = "data-persona"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.6",
    "django-cors-headers>=4.8.0",
    "djangorestframework>=3.16.1",
]
